backend/ai-service/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector Database & RAG
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Document Processing
pypdf2==3.0.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP & Networking
httpx==0.25.2
aiofiles==23.2.1

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Utilities
numpy==1.25.2
pandas==2.1.4

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging & Monitoring
structlog==23.2.0